feat: add module model
