+++ index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="styles/main.css" rel="stylesheet">
     <link href="styles/mobile.css" rel="stylesheet">
-    <script src="main.js" defer></script>
     <script src='https://cdn.jsdelivr.net/npm/big.js@6.0.0/big.min.js'></script>
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@100;400&display=swap" rel="stylesheet"> 
     <title>Calculator</title>
@@ -39,5 +38,8 @@
             <button data-equals id="equals" class="operation">=</button>            
         </div>
     </main>
+    
+    <script type="module" src="./main.js" crossorigin></script>
+    <script type="module" src="./script.js" crossorigin></script>
 </body>
 </html>
